start "Base" editor | add PulsarcDrawing
Added interface IEditor, abstract class EditorEngine, and other files that will be used with the Editor.

PulsarcDrawing is a static class with methods for creating Texture2D objects that contain shapes. Not using MonoGame.Extended.ShapeExtensions as that doesn't have what I'm looking for, and doesn't work well with Pulsarc's Drawables.
The reason for creating PulsarcDrawing is for future objects like "Beat Circles" that will show the user where beats are while editing.

Added a Lerp function for Color in PulsarcMath, in case we decide to expand PulsarcDrawing to draw gradients on shapes.

+++ Pulsarc/Beatmaps/TimingPoint.cs
@@ -9,9 +9,9 @@ namespace Pulsarc.Beatmaps
         // The Time this TimingPoint activates
         public int Time { get; private set; }
         // The BPM set by this TimingPoint
-        public int Bpm { get; private set; }
+        public double Bpm { get; private set; }
 
-        public TimingPoint(int time, int bpm)
+        public TimingPoint(int time, double bpm)
         {
             Time = time;
             Bpm = bpm;
